added ticker

+++ src/app/components/broker-details/broker-details.component.ts
@@ -2,6 +2,7 @@ import { Component, Input, OnInit } from '@angular/core';
 import { BrokerService } from 'src/app/services/broker.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Broker } from 'src/app/models/broker.model';
+import { FileImporter } from 'src/app/models/file-importer.enum';
 
 @Component({
   selector: 'app-broker-details',
@@ -11,7 +12,10 @@ import { Broker } from 'src/app/models/broker.model';
 export class BrokerDetailsComponent implements OnInit {
   @Input() viewMode = false;
   @Input() currentBroker: Broker = {
-    name: ''
+    brokerId: '',
+    name: '',
+    country: '',
+    fileImporter: FileImporter.Degiro
   };
   message = '';
   constructor(
@@ -19,49 +23,36 @@ export class BrokerDetailsComponent implements OnInit {
     private route: ActivatedRoute,
     private router: Router) { }
 
-    ngOnInit(): void {
-      if (!this.viewMode) {
-        this.message = '';
-        this.getBroker(this.route.snapshot.params["id"]);
-      }
-    }
-    getBroker(id: string): void {
-      this.brokerService.get(id)
-        .subscribe({
-          next: (data) => {
-            this.currentBroker = data;
-            console.log(data);
-          },
-          error: (e) => console.error(e)
-        });
-    }
-    updatePublished(status: boolean): void {
-      const data = {
-        name: this.currentBroker.name,
-      };
+  ngOnInit(): void {
+    if (!this.viewMode) {
       this.message = '';
-      this.brokerService.update(this.currentBroker.brokerId, data)
-        .subscribe({
-          next: (res) => {
-            console.log(res);
-            //this.currentBroker.published = status;
-            this.message = res.message ? res.message : 'The status was updated successfully!';
-          },
-          error: (e) => console.error(e)
-        });
+      this.getBroker(this.route.snapshot.params["id"]);
     }
-    updateTutorial(): void {
-      this.message = '';
-      this.brokerService.update(this.currentBroker.brokerId, this.currentBroker)
-        .subscribe({
-          next: (res) => {
-            console.log(res);
-            this.message = res.message ? res.message : 'This tutorial was updated successfully!';
-          },
-          error: (e) => console.error(e)
-        });
-    }
-    deleteTutorial(): void {
+  }
+  getBroker(id: string): void {
+    this.brokerService.get(id)
+      .subscribe({
+        next: (data) => {
+          this.currentBroker = data;
+          console.log(data);
+        },
+        error: (e) => console.error(e)
+      });
+  }
+  updateBroker(): void {
+    this.message = '';
+    this.brokerService.update(this.currentBroker)
+      .subscribe({
+        next: (res) => {
+          console.log(res);
+          this.message = 'This Broker was updated successfully!';
+        },
+        error: (e) => console.error(e)
+      });
+  }
+  deleteBroker(): void {
+    if(this.currentBroker.brokerId != undefined)
+    {
       this.brokerService.delete(this.currentBroker.brokerId)
         .subscribe({
           next: (res) => {
@@ -71,4 +62,7 @@ export class BrokerDetailsComponent implements OnInit {
           error: (e) => console.error(e)
         });
     }
+    else
+      throw "Invalid Broker"
+  }
 }
